# docs/swagger.yaml
basePath: /api
definitions:
...
  entity.ProjectRecruitmentHeaderStatus:
    enum:
    - DRAFT
    - SUBMITTED
    - APPROVED
    - REJECTED
    - CLOSE
    - IN PROGRESS
    - COMPLETED
    type: string
    x-enum-varnames:
    - PROJECT_RECRUITMENT_HEADER_STATUS_DRAFT
    - PROJECT_RECRUITMENT_HEADER_STATUS_SUBMITTED
    - PROJECT_RECRUITMENT_HEADER_STATUS_APPROVED
    - PROJECT_RECRUITMENT_HEADER_STATUS_REJECTED
    - PROJECT_RECRUITMENT_HEADER_STATUS_CLOSE
    - PROJECT_RECRUITMENT_HEADER_STATUS_IN_PROGRESS
    - PROJECT_RECRUITMENT_HEADER_STATUS_COMPLETED
...
  entity.ProjectRecruitmentType:
    enum:
    - MT_Management Trainee
    - PH_Professional Hire
    - NS_Non Staff to Staff
    type: string
    x-enum-varnames:
    - PROJECT_RECRUITMENT_TYPE_MT
    - PROJECT_RECRUITMENT_TYPE_PH
    - PROJECT_RECRUITMENT_TYPE_NS
...
  entity.TemplateActivityLineStatus:
    enum:
    - ACTIVE
    - INACTIVE
    type: string
    x-enum-varnames:
    - TEMPLATE_ACTIVITY_LINE_STATUS_ACTIVE
    - TEMPLATE_ACTIVITY_LINE_STATUS_INACTIVE
  entity.TemplateActivityStatus:
    enum:
    - ACTIVE
    - INACTIVE
    type: string
    x-enum-varnames:
    - TEMPLATE_ACTIVITY_STATUS_ACTIVE
    - TEMPLATE_ACTIVITY_STATUS_INACTIVE
...
  entity.TemplateQuestionFormType:
    enum:
    - ADMINISTRATIVE_SELECTION
    - TEST
    - INTERVIEW
    - FGD
    - FINAL_INTERVIEW
    - OFFERING_LETTER
    - CONTRACT_DOCUMENT
    - DOCUMENT_CHECKING
    - SURAT_IZIN_ORANG_TUA
    - KARYAWAN_TETAP
    - PKWT
    - PKWTT
    type: string
    x-enum-varnames:
    - TQ_FORM_TYPE_ADMINISTRATIVE_SELECTION
    - TQ_FORM_TYPE_TEST
    - TQ_FORM_TYPE_INTERVIEW
    - TQ_FORM_TYPE_FGD
    - TQ_FORM_TYPE_FINAL_INTERVIEW
    - TQ_FORM_TYPE_OFFERING_LETTER
    - TQ_FORM_TYPE_CONTRACT_DOCUMENT
    - TQ_FORM_TYPE_DOCUMENT_CHECKING
    - TQ_FORM_TYPE_SURAT_IZIN_ORTU
    - TQ_FORM_TYPE_KARYAWAN_TETAP
    - TQ_FORM_TYPE_PKWT
    - TQ_FORM_TYPE_PKWTT
  entity.TemplateQuestionStatus:
    enum:
    - ACTIVE
    - INACTIVE
    type: string
    x-enum-varnames:
    - TEMPLATE_QUESTION_STATUS_ACTIVE
    - TEMPLATE_QUESTION_STATUS_INACTIVE
...
  request.CreateProjectRecruitmentHeader:
    properties:
      description:
        type: string
      document_date:
        type: string
      document_number:
        type: string
      end_date:
        type: string
      name:
        type: string
      recruitment_type:
        type: string
      start_date:
        type: string
      status:
        type: string
      template_activity_id:
        type: string
    required:
    - document_date
    - document_number
    - end_date
    - name
    - recruitment_type
    - start_date
    - status
    - template_activity_id
    type: object
...
  request.UpdateProjectRecruitmentHeader:
    properties:
      description:
        type: string
      document_date:
        type: string
      document_number:
        type: string
      end_date:
        type: string
      id:
        type: string
      name:
        type: string
      recruitment_type:
        type: string
      start_date:
        type: string
      status:
        type: string
      template_activity_id:
        type: string
    required:
    - document_date
    - document_number
    - end_date
    - id
    - name
    - recruitment_type
    - start_date
    - status
    - template_activity_id
    type: object
...
  response.AnswerTypeResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  response.DocumentSetupResponse:
    properties:
      body:
        type: string
      created_at:
        type: string
      document_type:
        $ref: '#/definitions/response.DocumentTypeResponse'
      document_type_id:
        type: string
      footer:
        type: string
      header:
        type: string
      id:
        type: string
      recruitment_type:
        $ref: '#/definitions/entity.ProjectRecruitmentType'
      title:
        type: string
      updated_at:
        type: string
    type: object
  response.DocumentTypeResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  response.DocumentVerificationResponse:
    properties:
      format:
        type: string
      id:
        type: string
      name:
        type: string
      template_question:
        $ref: '#/definitions/response.TemplateQuestionResponse'
      template_question_id:
        type: string
    type: object
...
  response.ProjectPicResponse:
    properties:
      administrative_total:
        type: integer
      created_at:
        type: string
      employee_id:
        type: string
      employee_name:
        type: string
      id:
        type: string
      project_recruitment_line_id:
        type: string
      updated_at:
        type: string
    type: object
  response.ProjectRecruitmentHeaderResponse:
    properties:
      description:
        type: string
      document_date:
        type: string
      document_number:
        type: string
      end_date:
        type: string
      id:
        type: string
      name:
        type: string
      project_recruitment_lines:
        items:
          $ref: '#/definitions/response.ProjectRecruitmentLineResponse'
        type: array
      recruitment_type:
        $ref: '#/definitions/entity.ProjectRecruitmentType'
      start_date:
        type: string
      status:
        $ref: '#/definitions/entity.ProjectRecruitmentHeaderStatus'
      template_activity:
        $ref: '#/definitions/response.TemplateActivityResponse'
      template_activity_id:
        type: string
    type: object
  response.ProjectRecruitmentLineResponse:
    properties:
      created_at:
        type: string
      end_date:
        type: string
      id:
        type: string
      project_pics:
        items:
          $ref: '#/definitions/response.ProjectPicResponse'
        type: array
      project_recruitment_header_id:
        type: string
      start_date:
        type: string
      template_activity_id:
        type: string
      template_activity_line:
        $ref: '#/definitions/response.TemplateActivityLineResponse'
      updated_at:
        type: string
    type: object
  response.QuestionOptionResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      option_text:
        type: string
      question_id:
        type: string
      updated_at:
        type: string
    type: object
  response.QuestionResponse:
    properties:
      answer_type_id:
        type: string
      answer_types:
        $ref: '#/definitions/response.AnswerTypeResponse'
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      question_options:
        items:
          $ref: '#/definitions/response.QuestionOptionResponse'
        type: array
      template_question_id:
        type: string
      updated_at:
        type: string
    type: object
...
  response.TemplateActivityLineResponse:
    properties:
      color_hex_code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      question_template_id:
        type: string
      status:
        $ref: '#/definitions/entity.TemplateActivityLineStatus'
      template_activity:
        $ref: '#/definitions/response.TemplateActivityResponse'
      template_activity_id:
        type: string
      template_question:
        $ref: '#/definitions/response.TemplateQuestionResponse'
      updated_at:
        type: string
    type: object
  response.TemplateActivityResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      recruitment_type:
        $ref: '#/definitions/entity.ProjectRecruitmentType'
      status:
        $ref: '#/definitions/entity.TemplateActivityStatus'
      template_activity_lines:
        items:
          $ref: '#/definitions/response.TemplateActivityLineResponse'
        type: array
      updated_at:
        type: string
    type: object
  response.TemplateQuestionResponse:
    properties:
      description:
        type: string
      document_setup:
        $ref: '#/definitions/response.DocumentSetupResponse'
      document_setup_id:
        type: string
      document_verifications:
        items:
          $ref: '#/definitions/response.DocumentVerificationResponse'
        type: array
      duration:
        type: integer
      form_type:
        $ref: '#/definitions/entity.TemplateQuestionFormType'
      id:
        type: string
      name:
        type: string
      questions:
        items:
          $ref: '#/definitions/response.QuestionResponse'
        type: array
      status:
        $ref: '#/definitions/entity.TemplateQuestionStatus'
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Julong Recruitment API Docs
  version: "1.0"
paths:
...
  /api/project-recruitment-headers:
    get:
      consumes:
      - application/json
      description: Find all project recruitment headers
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page Size
        in: query
        name: page_size
        type: integer
      - description: Search
        in: query
        name: search
        type: string
      - description: Created At
        in: query
        name: created_at
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProjectRecruitmentHeaderResponse'
      security:
      - BearerAuth: []
      summary: Find all project recruitment headers
      tags:
      - Project Recruitment Headers
    post:
      consumes:
      - application/json
      description: Create project recruitment header
      parameters:
      - description: Create Project Recruitment Header
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/request.CreateProjectRecruitmentHeader'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.ProjectRecruitmentHeaderResponse'
      security:
      - BearerAuth: []
      summary: Create project recruitment header
      tags:
      - Project Recruitment Headers
  /api/project-recruitment-headers/{id}:
    delete:
      consumes:
      - application/json
      description: Delete project recruitment header
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete project recruitment header
      tags:
      - Project Recruitment Headers
    get:
      consumes:
      - application/json
      description: Find project recruitment header by id
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProjectRecruitmentHeaderResponse'
      security:
      - BearerAuth: []
      summary: Find project recruitment header by id
      tags:
      - Project Recruitment Headers
  /api/project-recruitment-headers/document-number:
    get:
      consumes:
      - application/json
      description: Generate document number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Generate document number
      tags:
      - Project Recruitment Headers
  /api/project-recruitment-headers/update:
    put:
      consumes:
      - application/json
      description: Update project recruitment header
      parameters:
      - description: Update Project Recruitment Header
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/request.UpdateProjectRecruitmentHeader'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProjectRecruitmentHeaderResponse'
      security:
      - BearerAuth: []
      summary: Update project recruitment header
      tags:
      - Project Recruitment Headers
...
securityDefinitions:
  BearerAuth:
    description: Bearer token for authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"